add testUsingAnnotationMockUser
