+++ src/main/java/org/forkingaround/adventuretime/config/SecurityConfig.java
@@ -47,10 +47,10 @@ public class SecurityConfig {
                                 .deleteCookies("ADVENTURER"))
                         .authorizeHttpRequests(auth -> auth
                                 //.requestMatchers(AntPathRequestMatcher.antMatcher("/h2-console/**")).permitAll()
-                                .requestMatchers(HttpMethod.GET, endpoint + "/login").permitAll()
                                 .requestMatchers(HttpMethod.POST, endpoint + "/register").permitAll()
                                 .requestMatchers(HttpMethod.GET, endpoint + "/home/*").permitAll()
-                                .requestMatchers(HttpMethod.GET, endpoint + "/event").hasRole("ADMIN")
+                                .requestMatchers(HttpMethod.GET, endpoint + "/login").hasAnyRole("USER","ADMIN")
+                                .requestMatchers(HttpMethod.GET, endpoint + "/event/*").hasRole("ADMIN")
                                 .requestMatchers(HttpMethod.GET, endpoint + "/dashboard").hasRole("ADMIN")
                                 .anyRequest().authenticated())
                         .userDetailsService(jpaUserDetailsService)
