--- src/main/java/org/forkingaround/adventuretime/controllers/DashboardController.java
@@ -1,16 +1,0 @@
-package org.forkingaround.adventuretime.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(path = "${api-endpoint}/dashboard")
-public class DashboardController {
-
-    @GetMapping( path = "")
-    public String index() {
-        return "Hello Spring Boot!!!";
-    }
-    
-}
