+++ src/main/resources/application-mysql.properties
@@ -3,7 +3,7 @@
 #Database - MySQL database in docker container
 spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/adventuretime
 spring.datasource.username=root
-spring.datasource.password=1234
+spring.datasource.password=12345
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
